# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"IUserToken": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"isAdmin": {
						"type": "boolean"
					}
				},
				"required": [
					"_id",
					"email",
					"name",
					"isAdmin"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUser": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"isAdmin": {
						"type": "boolean"
					}
				},
				"required": [
					"_id",
					"email",
					"name",
					"password",
					"isAdmin"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IParsingHotelResponse": {
				"properties": {
					"Hotel": {
						"type": "string"
					},
					"HotelID": {
						"type": "number",
						"format": "double"
					},
					"HotelServer": {
						"type": "string"
					},
					"PMS_ServerID": {
						"type": "number",
						"format": "double"
					},
					"ServerName": {
						"type": "string"
					}
				},
				"required": [
					"Hotel",
					"HotelID",
					"HotelServer",
					"PMS_ServerID",
					"ServerName"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			},
			"jwt-passport": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "hotels-api",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Vassil Shterev ",
			"email": "vsterev@gmail.com"
		}
	},
	"paths": {
		"/user/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"userData": {
											"$ref": "#/components/schemas/IUserToken"
										},
										"token": {
											"type": "string"
										}
									},
									"required": [
										"userData",
										"token"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/verify": {
			"get": {
				"operationId": "VerifyLogin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserToken"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [
					{
						"jwt-passport": []
					}
				],
				"parameters": []
			}
		},
		"/user/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IUser"
									},
									"type": "array"
								}
							}
						}
					},
					"403": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [
					{
						"jwt-passport": []
					}
				],
				"parameters": []
			}
		},
		"/user": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"403": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [
					{
						"jwt-passport": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"isAdmin": {
										"type": "boolean"
									},
									"password": {
										"type": "string"
									},
									"name": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"isAdmin",
									"password",
									"name",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"403": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [
					{
						"jwt-passport": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"_id": {
										"type": "string"
									}
								},
								"required": [
									"_id"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "EditUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"403": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [
					{
						"jwt-passport": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"isAdmin": {
										"type": "boolean"
									},
									"password": {
										"type": "string"
									},
									"name": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"_id": {
										"type": "string"
									}
								},
								"required": [
									"email",
									"_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/parser/hotels-compare": {
			"get": {
				"operationId": "HotelCompare",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"interlookHotels": {},
										"parserHotels": {
											"items": {
												"allOf": [
													{
														"$ref": "#/components/schemas/IParsingHotelResponse"
													},
													{
														"properties": {
															"mapped": {
																"type": "boolean"
															}
														},
														"required": [
															"mapped"
														],
														"type": "object"
													}
												]
											},
											"type": "array"
										}
									},
									"required": [
										"interlookHotels",
										"parserHotels"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"parser"
				],
				"security": [
					{
						"jwt-passport": []
					}
				],
				"parameters": []
			}
		},
		"/parser/hotel-props": {
			"post": {
				"operationId": "GetHotelProps",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"rooms": {
											"items": {},
											"type": "array"
										},
										"boards": {
											"items": {},
											"type": "array"
										}
									},
									"required": [
										"rooms",
										"boards"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"parser"
				],
				"security": [
					{
						"jwt-passport": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"parserCode": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"parserCode"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/parser/reports": {
			"post": {
				"operationId": "GetReports",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"parser"
				],
				"security": [
					{
						"jwt-passport": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"limit": {
										"type": "number",
										"format": "double"
									},
									"skip": {
										"type": "number",
										"format": "double"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}